==> notebooks/04 - doc2vec- training.ipynb <==
import pandas as pd
import pickle
import numpy as np

# Load the bar review dataset 
review = pd.read_pickle('../output/bar_reviews_cleaned_and_tokenized_SF.pickle')
print review.head(2)
print review.tail(2)

# Load the training users
user_set_training = pickle.load(open('../output/training_users.pickle', 'rb'))
# Make the active review set training only 
review = review[review.user_id.isin(user_set_training)]


from collections import OrderedDict
from itertools import chain


# n_reviews = 100 # all of them... 
# Flatten the reviews, so each review is just a single list of words.
reviews_merged_bus = OrderedDict()
business_set = list(set(review.business_id.values[:]))
for i_bus, bus_id in enumerate(business_set):
    if ((i_bus%5)==0):
        print '\r Fraction Processed',float(i_bus+1)/len(business_set),
    # This horrible line first collapses each review of a corresponding business into a list
    # of lists, and then collapses the list of sentences to a long list of words
    reviews_merged_bus[bus_id] = list(chain.from_iterable(chain.from_iterable( 
                                     review.cleaned_tokenized[review.business_id==bus_id] )) )
docs_bus = reviews_merged_bus.values()
print 
print len(docs_bus)

import gensim
from itertools import chain
import sys
sys.path.append('../vectorsearch/')
import nltk_helper
import doc2vec
from gensim.models.doc2vec import TaggedDocument
import pandas as pd
n_epochs = 10
n_docs = 10 # -1 for almost all of them...

# Generate the tagged document list. 

docs = [TaggedDocument(words, [business_set[index],])
                             for index, words in enumerate(docs_bus[:])]

print '\nFirst Doc: \n-----------------\n', docs[0]


from gensim.models import doc2vec

model = doc2vec.Doc2Vec(min_count=4, window=5, size=200, sample=1e-4, negative=10, workers=12)
# Build the vocab from list of sentences.
model.build_vocab(docs) 



from random import shuffle

for epoch in range(10):
    print '\rTraining Epoch %i, alpha %1.4f'%(epoch+1, model.alpha),
    #model.train(np.random.permutation(docs))
    shuffle(docs)
    model.train(docs)
    model.alpha -= 0.001 # decrease the learning rate
    model.min_alpha = model.alpha # fix the learning rate, no decay

#model.init_sims(replace=True)    
# # Normalize the word vectors.
# vec_norms = np.sqrt(np.sum(model.syn0**2, axis=1))
# model.syn0 = (model.syn0/vec_norms[:, numpy.newaxis])
# # Normalize the doc vectors.
# vec_norms = np.sqrt(np.sum(model.docvecs.doctag_syn0**2, axis=1))
# model.docvecs.doctag_syn0 = (model.docvecs.doctag_syn0/vec_norms[:, numpy.newaxis])

model.save('../output/doc2vec_bars_200_neg_10_win_5.model')

# Can find similar documents..
print model.docvecs.most_similar(positive=['KUinHkKyGhznElgIzx0yIw']), '\n'

# Can find similar words...Re: Dream Companies and contact from recruiters
print model.most_similar(positive=['beer']), '\n'

# Can find documents that are most similar to keywords.... 
print model.docvecs.most_similar(positive=[model['beer'], model['music']]), '\n'

# Can find words that are most common in documents
print review.text[review.review_id=='KUinHkKyGhznElgIzx0yIw'].values
print model.most_similar(positive=[model.docvecs['KUinHkKyGhznElgIzx0yIw']]), '\n'
